refactor(backend): extract session handlers
